Refactor: 모임 날짜 컴포넌트 분리
