+++ modules/post/components/voteDateMessage/style.ts
