+++ modules/post/components/voteDateMessage/type.ts
